local tests pass, add a single conftest to try to reduce overhead and number of github API calls
