+++ tests/test_IoU.py
@@ -1,19 +1,14 @@
 #Test IoU
-from .conftest import download_release
 from deepforest import IoU
-from deepforest import main
 from deepforest import get_data
 from deepforest import visualize
 
 import os
-import pytest
 import shapely
 import geopandas as gpd
 import pandas as pd
 
-def test_compute_IoU(download_release):
-    m = main.deepforest()
-    m.use_release(check_release=False)
+def test_compute_IoU(m):
     csv_file = get_data("OSBS_029.csv")
     predictions = m.predict_file(csv_file=csv_file, root_dir=os.path.dirname(csv_file))
     ground_truth = pd.read_csv(csv_file)
