+++ tests/test_callbacks.py
@@ -3,25 +3,7 @@ from deepforest import main
 from deepforest import callbacks
 import glob
 import os
-import pytest
-
 from deepforest import get_data
-from .conftest import download_release
-
-@pytest.fixture()
-def m(download_release):
-    m = main.deepforest()
-    m.config["train"]["csv_file"] = get_data("example.csv") 
-    m.config["train"]["root_dir"] = os.path.dirname(get_data("example.csv"))
-    m.config["train"]["fast_dev_run"] = False
-    m.config["batch_size"] = 2
-       
-    m.config["validation"]["csv_file"] = get_data("example.csv") 
-    m.config["validation"]["root_dir"] = os.path.dirname(get_data("example.csv"))
-    
-    m.use_release(check_release=False)
-    
-    return m
 
 def test_log_images(m, tmpdir):
     im_callback = callbacks.images_callback(csv_file=m.config["validation"]["csv_file"], root_dir=m.config["validation"]["root_dir"], savedir=tmpdir)
