+++ tests/test_environment.py
@@ -1,6 +1,8 @@
 #test environment
 
-
 def test_environment():
     from deepforest.main import deepforest
     deepforest_class = deepforest()
+
+def test_main_import():
+    from deepforest import main
