+++ tests/test_main.py
@@ -19,48 +19,10 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import Callback
 from PIL import Image
 
-#Import release model from global script to avoid thrasing github during testing. Just download once.
-from .conftest import download_release
-
 @pytest.fixture()
-def two_class_m():
-    m = main.deepforest(num_classes=2,label_dict={"Alive":0,"Dead":1})
-    m.config["train"]["csv_file"] = get_data("testfile_multi.csv") 
-    m.config["train"]["root_dir"] = os.path.dirname(get_data("testfile_multi.csv"))
-    m.config["train"]["fast_dev_run"] = True
-    m.config["batch_size"] = 2
-        
-    m.config["validation"]["csv_file"] = get_data("testfile_multi.csv") 
-    m.config["validation"]["root_dir"] = os.path.dirname(get_data("testfile_multi.csv"))
-    m.config["validation"]["val_accuracy_interval"] = 1
-
-    m.create_trainer()
-    
-    return m
-
-@pytest.fixture()
-def m(download_release):
-    m = main.deepforest()
-    m.config["train"]["csv_file"] = get_data("example.csv") 
-    m.config["train"]["root_dir"] = os.path.dirname(get_data("example.csv"))
-    m.config["train"]["fast_dev_run"] = True
-    m.config["batch_size"] = 2
-        
-    m.config["validation"]["csv_file"] = get_data("example.csv") 
-    m.config["validation"]["root_dir"] = os.path.dirname(get_data("example.csv"))
-    m.config["workers"] = 0 
-    m.config["validation"]["val_accuracy_interval"] = 1
-    m.config["train"]["epochs"] = 2
-    
-    m.create_trainer()
-    m.use_release(check_release=False)
-    
-    return m
-
 def big_file():
     tmpdir = tempfile.gettempdir()
     csv_file = get_data("OSBS_029.csv")
-    image_path = get_data("OSBS_029.png")
     df = pd.read_csv(csv_file)    
     
     big_frame = []
@@ -75,9 +37,6 @@ def big_file():
     big_frame.to_csv("{}/annotations.csv".format(tmpdir))    
     
     return "{}/annotations.csv".format(tmpdir)
-    
-def test_main():
-    from deepforest import main
 
 def test_use_bird_release(m):
     imgpath = get_data("AWPE Pigeon Lake 2020 DJI_0005.JPG")
